% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{append_to_data}
\alias{append_to_data}
\title{append_to_data}
\usage{
append_to_data(
  dataframe,
  Date = as.character(format(Sys.Date(), "\%d-\%m-\%Y")),
  Hour = substr(Sys.time(), 12, 16),
  ...
)
}
\arguments{
\item{dataframe}{The dataframe to append data to}

\item{Date}{The date at which the event happened. Format dd-mm-yyy}

\item{Hour}{The hour at which the event happened. Format hh:mm}

\item{...}{the data to append. It has to be named and to be related to the different names}
}
\value{
new data frame with new rows
}
\description{
append_to_data
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
append_to_data(dataframe = dummy_data, Lactation_Left = TRUE, Temperature = 37.2)

}
