% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_new_data}
\alias{create_new_data}
\title{create_new_data}
\usage{
create_new_data(firstname = NULL, surname = NULL, previous_data = NULL)
}
\arguments{
\item{firstname}{First name of your kid.
Defaut is NULL leading to John/Jane}

\item{surname}{Surname of your kid.
Defaut is NULLL leading to Doh.}

\item{previous_data}{Prexisting data. Must be NULL or a csv leading to a dataframe
Defaut is NULL leading to an empty data frame}
}
\value{
dataframe A dataframe with names as attributes and containing several columns :
   - Weight (numeric)
   - Temperature (numeric)
   - Date (datetime)
   - Hour (datetime)
   - Mother milk (numeric)
   - Powder milk (numeric)
   - Lactation left (boolean)
   - Lactation right (boolean)
   - Vomit (integer)
   - Urin (integer)
   - Poop (integer)
}
\description{
create_new_data
}
\examples{
create_new_data(firstname = "Test", surname = "Retest")

}
