% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{lactation_graph}
\alias{lactation_graph}
\title{lactation_graph}
\usage{
lactation_graph(dataframe = NULL, granularity = "Hour")
}
\arguments{
\item{dataframe}{The dataframe containing the data you want to use}

\item{granularity}{The granularity you want to use. For the moment it is hour or day.}
}
\value{
graphic plotly object
}
\description{
Function to plot lactation vs time
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
lactation_graph(dataframe = dummy_data)

}
