% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{mean_df}
\alias{mean_df}
\title{mean_df}
\usage{
mean_df(dataframe = NULL, granularity_mean = "Day", birthdate = NULL)
}
\arguments{
\item{dataframe}{The dataframe from which you want to calculate the mean. It must have a Time column and the numeric column from which you want to have the mean.}

\item{granularity_mean}{The granularity at which you want to calculate the mean. Must be Day or Week. Defaut is Day}

\item{birthdate}{The birthdate of the child. If NULL we assume that this is the first day of the dataframe. Defaut is NULL}
}
\value{
mean_df the new dataframe containing the mean values
}
\description{
function returning the mean value of a data frame
}
