% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{milk_feeding_graph}
\alias{milk_feeding_graph}
\title{milk_feeding_graph}
\usage{
milk_feeding_graph(
  dataframe = NULL,
  granularity = "Hour",
  mean_plot = FALSE,
  birthdate = NULL
)
}
\arguments{
\item{dataframe}{The dataframe containing the data you want to use}

\item{granularity}{The granularity you want to use. For the moment it is hour or day.}

\item{mean_plot}{Boolean to plot mean with granularity one step higher than the asked one.}

\item{birthdate}{The birthdate of the child. If NULL we assume that this is the first day of the dataframe. Defaut is NULL}
}
\value{
graphic plotly object
}
\description{
Graph to survey milk feeding versus time
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
milk_feeding_graph(dataframe = dummy_data)

}
