% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_data}
\alias{save_data}
\title{save_data}
\usage{
save_data(dataframe = NULL, filename = NULL, path = NULL)
}
\arguments{
\item{dataframe}{The dataframe to be saved}

\item{filename}{The filename to save the file in}

\item{path}{The path to save the file in}
}
\value{
csv file
}
\description{
function to save the data
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
save_data(dataframe = dummy_data, filename = "test")
}
