% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{temperature_graph}
\alias{temperature_graph}
\title{temperature_graph}
\usage{
temperature_graph(dataframe = NULL, birthdate = NULL)
}
\arguments{
\item{dataframe}{The dataframe you want to analyze}

\item{birthdate}{The birthdate of the child. If NULL we assume that this is the first day of the dataframe. Defaut is NULL}
}
\value{
plotly graphic object
}
\description{
Function to plot the temperature versus time
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
temperature_graph(dataframe = dummy_data)

}
