% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{weight_graph}
\alias{weight_graph}
\title{weight_graph}
\usage{
weight_graph(dataframe = NULL, birthdate = NULL)
}
\arguments{
\item{dataframe}{The dataframe contaning the dat you want to use}

\item{birthdate}{The birthdate of the child. If NULL we assume that this is the first day of the dataframe. Defaut is NULL}
}
\value{
graphic plotly object
}
\description{
function to plot the weight vs date
}
\examples{
dummy_data <- utils::read.csv(file.path(system.file("extdata", package = "NewBoRn"),
"dummy_data.csv"))
dummy_data <- dplyr::select(.data = dummy_data, -X)
weight_graph(dataframe = dummy_data)

}
